--- src/ToDo.txt
@@ -1,1 +1,0 @@
-REAFACTOR -> wspólne funkcje statyczne dla mastera i slave przerzucić do jednego pliku common w internall
