+++ src/common/modbus_PDU.c
@@ -8,7 +8,7 @@
  * @copyright Copyright (c) 2024
  * 
  */
-#include "modbus_PDU_common.h"
+#include "modbus_PDU.h"
 #include "modbus_data.h"
 
 modbus_byte_count_t get_coil_din_byte_count(modbus_data_qty_t coil_qty)
