+++ src/slave_internall/modbus_slave_PDU.c
@@ -9,7 +9,7 @@
  */
 
 #include "modbus_slave_PDU.h"
-#include "modbus_PDU_common.h"
+#include "modbus_PDU.h"
 #include "modbus_data.h"
 #include "buf_rw.h"
 // #include "modbus_type.h"
